Updated generally

+++ WakeMeUp/UIImage+FlatUI.m
@@ -137,7 +137,7 @@ static CGFloat edgeSizeFromCornerRadius(CGFloat cornerRadius) {
     ctrlPoint.y += radius;
     mPoint.x += radius;
     mPoint.y += radius;
-    if (radius > 0) [path addArcWithCenter:ctrlPoint radius:radius startAngle:(float)M_PI + (float)M_PI_2 endAngle:0 clockwise:YES];
+    if (radius > 0) [path addArcWithCenter:ctrlPoint radius:radius startAngle:(double)M_PI + (double)M_PI_2 endAngle:0 clockwise:YES];
     
     mPoint.y = CGRectGetMaxY(rect) - radius;
     [path addLineToPoint:mPoint];
@@ -146,7 +146,7 @@ static CGFloat edgeSizeFromCornerRadius(CGFloat cornerRadius) {
     mPoint.y += radius;
     mPoint.x -= radius;
     ctrlPoint.x -= radius;
-    if (radius > 0) [path addArcWithCenter:ctrlPoint radius:radius startAngle:0 endAngle:(float)M_PI_2 clockwise:YES];
+    if (radius > 0) [path addArcWithCenter:ctrlPoint radius:radius startAngle:0 endAngle:(double)M_PI_2 clockwise:YES];
     
     mPoint.x = rect.origin.x + (10.0f);
     [path addLineToPoint:mPoint];
